requirements.txt
pyfiglet
termcolor
PyYAML
Flask
rich
regex
rule-engine
jinja2
flask
google-play-scraper
requests
beautifulsoup4
threading
